test: ensure database is cleaned between tests
