+++ Gemfile
@@ -2,6 +2,7 @@ source 'https://rubygems.org'
 
 group :development, :test do
   gem 'activerecord', '~> 6.0'
+  gem 'database_cleaner-active_record'
   gem 'pg', '~> 1.2'
   gem 'rspec', '~> 3.10'
 end
